fix(aarch64): correct put_movk_reg_imm parameter types
The function signature expects (reg, imm: u16, shift: u32), not (reg, imm: u32, shift: u8).
This was revealed after merging upstream which had updated bindings.

+++ frida-gum/src/instruction_writer/aarch64/writer.rs
@@ -313,7 +313,7 @@ impl Aarch64InstructionWriter {
     /// Moves a 16-bit immediate into a register while keeping other bits unchanged.
     /// Used for building 64-bit constants across multiple instructions.
     /// Added in Frida 17.16.1.
-    pub fn put_movk_reg_imm(&self, reg: Aarch64Register, imm: u32, shift: u8) -> bool {
+    pub fn put_movk_reg_imm(&self, reg: Aarch64Register, imm: u16, shift: u32) -> bool {
         unsafe {
             gum_sys::gum_arm64_writer_put_movk_reg_imm(self.writer, reg as u32, imm, shift) != 0
         }
